.emacs: add debian files back .sawfishrc: paste from the phone, and undo from qtcreator bin/.gitx: git export rev file bin/Linux/config/pkgs/python-opencc: add opencc bin/adb-get-logs: add more logs bin/aof: fix-deep-dirs bin/bhj-help-it: add more search items than methods (props and enums) bin/emacs-get-android-clipboard: paste into emacs from android bin/gbkunzip: fix-deep-dirs bin/git-remote-url: for git svn clone bin/putclip-android: getclip-android bin/smbcp: push a directory etc/adb-sync-data-exclude: sort and add gcode/SetClip/src/com/bhj/setclip/PutClipService.java: add getclip function

+++ src/com/bhj/setclip/PutClipService.java
@@ -8,6 +8,7 @@ import android.os.IBinder;
 import android.widget.Toast;
 import java.io.File;
 import java.io.FileReader;
+import java.io.FileWriter;
 
 public class PutClipService extends Service {
     @Override
@@ -18,15 +19,25 @@ public class PutClipService extends Service {
     @Override
     public int onStartCommand(Intent intent,  int flags,  int startId)  {
         try {
-            FileReader f = new FileReader(new File("/sdcard/putclip.txt"));
-            char[] buffer = new char[1024];
-            int n = f.read(buffer);
-            String str = new String(buffer, 0, n);
-            ClipboardManager mClipboard;
-            mClipboard = (ClipboardManager)getSystemService(CLIPBOARD_SERVICE);
-            mClipboard.setPrimaryClip(ClipData.newPlainText("Styled Text", str));
-            File putclip = new File("/sdcard/putclip.txt");
-            putclip.delete();
+            if (intent.getIntExtra("getclip", 0) == 1) {
+                FileWriter f = new FileWriter(new File("/sdcard/putclip.txt.1"));
+                ClipboardManager mClipboard = (ClipboardManager)getSystemService(CLIPBOARD_SERVICE);
+                String str = mClipboard.getPrimaryClip().getItemAt(0).getText().toString();
+                f.write(str);
+                f.close();
+                File txt = new File("/sdcard/putclip.txt.1");
+                txt.renameTo(new File("/sdcard/putclip.txt"));
+            } else {
+                FileReader f = new FileReader(new File("/sdcard/putclip.txt"));
+                char[] buffer = new char[1024 * 1024];
+                int n = f.read(buffer);
+                String str = new String(buffer, 0, n);
+                ClipboardManager mClipboard;
+                mClipboard = (ClipboardManager)getSystemService(CLIPBOARD_SERVICE);
+                mClipboard.setPrimaryClip(ClipData.newPlainText("Styled Text", str));
+                File putclip = new File("/sdcard/putclip.txt");
+                putclip.delete();
+            }
         } catch (Exception e) {
             Toast.makeText(this, "Something went wrong in putclip: " + e.getMessage(), Toast.LENGTH_SHORT).show();
         }
